+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Bump `@vtex/ads-react` from `0.3.1` to `0.5.4`
+
 ## [3.147.0] - 2026-05-21
 
 ### Changed